# Conflicts:
#	cpct-common/src/main/java/com/zjtelcom/cpct/dao/channel/ObjMktCampaignRelMapper.java
#	cpct-common/src/main/java/com/zjtelcom/cpct/domain/channel/ObjMktCampaignRel.java
#	cpct-common/src/main/java/com/zjtelcom/cpct/dto/channel/CampaignInstVO.java
#	cpct-dubbo-server/src/main/java/com/zjtelcom/cpct/dubbo/service/impl/MktCampaignSyncApiServiceImpl.java
#	cpct-web/src/main/java/com/zjtelcom/cpct/Application.java

+++ cpct-service/src/main/java/com/zjtelcom/cpct/service/impl/campaign/MktCamChlConfServiceImpl.java
@@ -110,12 +110,14 @@ public class MktCamChlConfServiceImpl extends BaseService implements MktCamChlCo
             mktCamChlConfMap.put("resultCode", CommonConstant.CODE_SUCCESS);
             mktCamChlConfMap.put("resultMsg", ErrorCode.SAVE_CAM_CHL_CONF_SUCCESS.getErrorMsg());
             mktCamChlConfMap.put("evtContactConfId", evtContactConfId);
+            mktCamChlConfMap.put("evtContactConfName",mktCamChlConfDO.getEvtContactConfName());
             MktCamChlConfDetail mktCamChlConfDetailNew = BeanUtil.create(mktCamChlConfDO, new MktCamChlConfDetail());
             // 添加属性
             List<MktCamChlConfAttr> mktCamChlConfAttrList = mktCamChlConfDetail.getMktCamChlConfAttrList();
             List<MktCamChlConfAttrDO> mktCamChlConfAttrDOList = new ArrayList<>();
             for (MktCamChlConfAttr mktCamChlConfAttr : mktCamChlConfAttrList) {
                 MktCamChlConfAttrDO mktCamChlConfAttrDO = BeanUtil.create(mktCamChlConfAttr, new MktCamChlConfAttrDO());
+                mktCamChlConfAttrDO.setStatusCd(StatusCode.STATUS_CODE_EFFECTIVE.getStatusCode());
                 mktCamChlConfAttrDO.setEvtContactConfId(mktCamChlConfDO.getEvtContactConfId());
                 if (mktCamChlConfAttr.getAttrId().equals(ConfAttrEnum.RULE.getArrId())) {
                     mktCamChlConfAttrDO.setAttrValue(evtContactConfId.toString());
@@ -126,8 +128,9 @@ public class MktCamChlConfServiceImpl extends BaseService implements MktCamChlCo
                 //mktCamChlConfAttrMapper.insert(mktCamChlConfAttrDO);
                 mktCamChlConfAttrDOList.add(mktCamChlConfAttrDO);
             }
-            mktCamChlConfAttrMapper.insertBatch(mktCamChlConfAttrDOList);
-
+            if (mktCamChlConfAttrDOList.size()>0){
+                mktCamChlConfAttrMapper.insertBatch(mktCamChlConfAttrDOList);
+            }
             List<MktCamChlConfAttr> mktCamChlConfAttrNewList = new ArrayList<>();
             for (MktCamChlConfAttrDO mktCamChlConfAttrDO : mktCamChlConfAttrDOList) {
                 MktCamChlConfAttr mktCamChlConfAttr = BeanUtil.create(mktCamChlConfAttrDO, new MktCamChlConfAttr());
@@ -136,6 +139,9 @@ public class MktCamChlConfServiceImpl extends BaseService implements MktCamChlCo
             mktCamChlConfDetailNew.setMktCamChlConfAttrList(mktCamChlConfAttrNewList);
             // 将推送渠道缓存到redis
             redisUtils.set("MktCamChlConfDetail_" + evtContactConfId, mktCamChlConfDetailNew);
+
+//            ruleInsert(1L, "{\"type\":1,\"listData\":[{\"id\":177,\"name\":\"自有标签测试\",\"operatorList\":[{\"operName\":\"小于\",\"operValue\":2000},{\"operName\":\"等于\",\"operValue\":3000},{\"operName\":\"大于\",\"operValue\":1000},{\"operName\":\"不等于\",\"operValue\":4000},{\"operName\":\"小于等于\",\"operValue\":6000},{\"operName\":\"大于等于\",\"operValue\":5000},{\"operName\":\"包含\",\"operValue\":7000},{\"operName\":\"区间于\",\"operValue\":7200}],\"operType\":3000,\"valueList\":[{\"injectionLabelId\":177,\"labelValue\":\"是\",\"valueName\":\"是\"}],\"conditionType\":\"4\",\"conditionFour\":\"nhnh\",\"content\":\"nhnh\"}],\"ruleChildren\":{\"type\":2,\"listData\":[{\"id\":178,\"name\":\"话费余额\",\"operatorList\":[{\"operName\":\"小于\",\"operValue\":2000},{\"operName\":\"等于\",\"operValue\":3000},{\"operName\":\"大于\",\"operValue\":1000},{\"operName\":\"不等于\",\"operValue\":4000},{\"operName\":\"小于等于\",\"operValue\":6000},{\"operName\":\"大于等于\",\"operValue\":5000},{\"operName\":\"包含\",\"operValue\":7000},{\"operName\":\"区间于\",\"operValue\":7200}],\"operType\":3000,\"valueList\":[{\"injectionLabelId\":178,\"labelValue\":\"是\",\"valueName\":\"是\"}],\"conditionType\":\"4\",\"conditionFour\":\"gnghn\",\"content\":\"gnghn\"}]}}");
+
         } catch (Exception e) {
             logger.error("[op:MktCamChlConfServiceImpl] fail to save MktCamChlConf = {}", mktCamChlConfDO, e);
             mktCamChlConfMap.put("resultCode", CommonConstant.CODE_FAIL);
@@ -173,11 +179,11 @@ public class MktCamChlConfServiceImpl extends BaseService implements MktCamChlCo
                 MktCamChlConfAttrDO mktCamChlConfAttrDO = new MktCamChlConfAttrDO();
                 CopyPropertiesUtil.copyBean2Bean(mktCamChlConfAttrDO, mktCamChlConfAttr);
                 mktCamChlConfAttrDO.setEvtContactConfId(mktCamChlConfDO.getEvtContactConfId());
-                if (mktCamChlConfAttr.getAttrId().equals(ConfAttrEnum.RULE.getArrId())) {
+                if (mktCamChlConfAttr.getAttrId()!=null && mktCamChlConfAttr.getAttrId().equals(ConfAttrEnum.RULE.getArrId())) {
                     mktCamChlConfAttrDO.setAttrValue(evtContactConfId.toString());
                     //删除旧的关联规则 todo 静态
                     mktVerbalConditionMapper.deleteByVerbalId("1", evtContactConfId);
-                    //保存新的规则
+                    //保存新的规则 j
                     String params = mktCamChlConfAttr.getAttrValue();
                     ruleInsert(evtContactConfId, params);
                 }
@@ -208,6 +214,13 @@ public class MktCamChlConfServiceImpl extends BaseService implements MktCamChlCo
             MktCamChlConfDO mktCamChlConfDO = mktCamChlConfMapper.selectByPrimaryKey(evtContactConfId);
             List<MktCamChlConfAttrDO> mktCamChlConfAttrDOList = mktCamChlConfAttrMapper.selectByEvtContactConfId(evtContactConfId);
             CopyPropertiesUtil.copyBean2Bean(mktCamChlConfDetail, mktCamChlConfDO);
+            // 通过查询结果与推送渠道的关系，判断是否为二次协同
+            MktCamChlResultConfRelDO mktCamChlResultConfRelDO = mktCamChlResultConfRelMapper.selectByConfId(evtContactConfId);
+            if (mktCamChlResultConfRelDO != null) {
+                mktCamChlConfDetail.setIsSecondCoop("1");
+            } else {
+                mktCamChlConfDetail.setIsSecondCoop("0");
+            }
             List<MktCamChlConfAttr> mktCamChlConfAttrList = new ArrayList<>();
             for (MktCamChlConfAttrDO mktCamChlConfAttrDO : mktCamChlConfAttrDOList) {
                 MktCamChlConfAttr mktCamChlConfAttr = new MktCamChlConfAttr();
@@ -300,9 +313,11 @@ public class MktCamChlConfServiceImpl extends BaseService implements MktCamChlCo
      */
     @Override
     public Map<String, Object> deleteMktCamChlConf(Long evtContactConfId, Long ruleId) {
-        mktCamChlConfMapper.deleteByPrimaryKey(evtContactConfId);
-        mktCamChlConfAttrMapper.deleteByEvtContactConfId(evtContactConfId);
-        //判断是否是结果下的
+        Map<String, Object> mktCamChlConfMap = null;
+        try {
+            mktCamChlConfMapper.deleteByPrimaryKey(evtContactConfId);
+            mktCamChlConfAttrMapper.deleteByEvtContactConfId(evtContactConfId);
+            //判断是否是结果下的
 /*        MktCamChlResultConfRelDO mktCamChlResultConfRelDO = mktCamChlResultConfRelMapper.selectByConfId(evtContactConfId);
         if (mktCamChlResultConfRelDO != null) {
             mktCamChlResultConfRelMapper.deleteByPrimaryKey(mktCamChlResultConfRelDO.getMktCamChlResultConfRelId());
@@ -325,14 +340,19 @@ public class MktCamChlConfServiceImpl extends BaseService implements MktCamChlCo
         }*/
 
 
-        //删除旧的关联规则 todo 静态
-        mktVerbalConditionMapper.deleteByVerbalId("1", evtContactConfId);
+            //删除旧的关联规则 todo 静态
+            mktVerbalConditionMapper.deleteByVerbalId("1", evtContactConfId);
 
-        // 将推送渠道缓存到redis
-        redisUtils.remove("MktCamChlConfDetail_" + evtContactConfId);
-        Map<String, Object> mktCamChlConfMap = new HashMap<>();
-        mktCamChlConfMap.put("resultCode", CommonConstant.CODE_SUCCESS);
-        mktCamChlConfMap.put("resultMsg", ErrorCode.DELETE_CAM_CHL_CONF_SUCCESS.getErrorMsg());
+            // 将推送渠道缓存到redis
+            redisUtils.remove("MktCamChlConfDetail_" + evtContactConfId);
+            mktCamChlConfMap = new HashMap<>();
+            mktCamChlConfMap.put("resultCode", CommonConstant.CODE_SUCCESS);
+            mktCamChlConfMap.put("resultMsg", ErrorCode.DELETE_CAM_CHL_CONF_SUCCESS.getErrorMsg());
+        } catch (Exception e) {
+            logger.error("[op:MktCamChlConfServiceImpl] fail to delete mktCamChlConfDetailList evtContactConfId = {}", evtContactConfId, e);
+            mktCamChlConfMap.put("resultCode", CommonConstant.CODE_FAIL);
+            mktCamChlConfMap.put("resultMsg", ErrorCode.DELETE_CAM_CHL_CONF_FAILURE.getErrorMsg());
+        }
         return mktCamChlConfMap;
     }
 
@@ -446,7 +466,7 @@ public class MktCamChlConfServiceImpl extends BaseService implements MktCamChlCo
         List<MktVerbalCondition> labels = new ArrayList<>(); //标签因子
         List<MktVerbalCondition> expressions = new ArrayList<>(); //表达式
 
-        //分类
+        //将标签和表达式分类
         for (MktVerbalCondition mktVerbalCondition : mktVerbalConditions) {
             if ("1000".equals(mktVerbalCondition.getLeftParamType())) {
                 labels.add(mktVerbalCondition);
@@ -666,7 +686,9 @@ public class MktCamChlConfServiceImpl extends BaseService implements MktCamChlCo
             // 查询痛痒点话术列表
             verbalService.copyVerbal(parentEvtContactConfId, childEvtContactConfId);
             // 查询脚本
-            camScriptService.copyCamScript(parentEvtContactConfId, scriptDesc, childEvtContactConfId);
+            Map<String, Object> map = camScriptService.copyCamScript(parentEvtContactConfId, scriptDesc, childEvtContactConfId);
+            CamScript newScript = (CamScript) map.get("resultMsg");
+            mktCamChlConfDetailNew.setCamScript(newScript);
             mktCamChlConfMap.put("resultCode", CommonConstant.CODE_SUCCESS);
             mktCamChlConfMap.put("resultMsg", ErrorCode.SAVE_CAM_CHL_CONF_SUCCESS.getErrorMsg());
             mktCamChlConfMap.put("mktCamChlConfDetail", mktCamChlConfDetailNew);
